requirements/local.txt
# uninstall all the pip packages, if needed
#       pip freeze | xargs pip uninstall -y

-r base.txt # includes the base.txt requirements file

django-debug-toolbar==1.8